Refactor Store Controller
