+++ src/main/java/com/ntoutakeout/backend/controller/v1/AuthController.java
@@ -1,0 +1,11 @@
+package com.ntoutakeout.backend.controller.v1;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/api/v1/auth")
+@Slf4j
+public class AuthController {
+}
