+++ src/test/java/com/ntoutakeout/backend/NtouTakeoutBackendApplicationTests.java
@@ -1,24 +1,8 @@
 package com.ntoutakeout.backend;
 
-import com.ntoutakeout.backend.controller.StoreController;
-import com.ntoutakeout.backend.entity.Menu;
-import com.ntoutakeout.backend.entity.Review;
-import com.ntoutakeout.backend.entity.Store;
-import com.ntoutakeout.backend.service.StoreService;
-import org.junit.jupiter.api.Test;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.MockitoAnnotations;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.http.ResponseEntity;
-
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.mockito.Mockito.when;
 
 @SpringBootTest
 class NtouTakeoutBackendApplicationTests {
